requirements.txt
Django==1.11.6
django-oauth-plus==2.2.9
djangorestframework==3.7.1
djangorestframework-oauth==1.1.0
